add removed flag where necessary for api response
